Update CourseView to be the same as LectureView
There is no need for people to go to CourseView, then LectureView. This is an extra unnecessary click for new users wanting to check the content, so now we just show the lecture view straight view defaulting to first lecture
- Also make the lecture view put the video on top in mobile view
